# admin/doc/swagger.yaml
info:
  contact: {}
paths:
...
  /config/api/route:
    delete:
      parameters:
      - description: Route binding name
        in: query
        name: name
        required: true
        type: string
      - description: Draft key mod revision
        in: query
        name: expectedRevision
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: delete and publish one API route binding
      tags:
      - Config
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: create API route binding draft
      tags:
      - Config
    put:
      consumes:
      - application/json
      parameters:
      - description: Original route binding name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: modify API route binding draft
      tags:
      - Config
  /config/api/route/detail:
    get:
      parameters:
      - description: Route binding name
        in: query
        name: name
        required: true
        type: string
      - description: draft or published
        in: query
        name: scope
        type: string
      - description: 1 for draft, 0 for published
        in: query
        name: unpublished
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: get API route binding detail
      tags:
      - Config
  /config/api/route/diff:
    get:
      parameters:
      - description: Route binding name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: get API route binding diff
      tags:
      - Config
  /config/api/route/list:
    get:
      parameters:
      - description: draft or published
        in: query
        name: scope
        type: string
      - description: 1 for draft, 0 for published
        in: query
        name: unpublished
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: get API route binding list
      tags:
      - Config
...
  /config/api/route/publish:
    put:
      parameters:
      - description: Route binding name
        in: query
        name: name
        required: true
        type: string
      - description: Draft route key mod revision
        in: query
        name: expectedRevision
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: atomically publish one API route binding
      tags:
      - Config
...
  /config/api/route/status:
    get:
      parameters:
      - description: Route binding name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: get API route binding status
      tags:
      - Config
...
swagger: "2.0"